[core] Split buildApi.ts
